+++ hug/__init__.py
@@ -29,12 +29,12 @@ CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFT
 OTHER DEALINGS IN THE SOFTWARE.
 
 """
-from hug import (authentication, directives, documentation, exceptions, format, input_format, middleware, output_format,
-                 run, test, types)
+from hug import (authentication, directives, documentation, exceptions, format,
+                 input_format, middleware, output_format, run, test, types)
 from hug._version import current
-from hug.decorators import (call, cli, connect, default_input_format, default_output_format, delete, directive,
-                            extend_api, get, head, not_found, options, patch, post, put, request_middleware,
-                            response_middleware, trace)
+from hug.decorators import (call, cli, connect, default_input_format, default_output_format,
+                            delete, directive, extend_api, get, head, not_found, options,
+                            patch, post, put, request_middleware, response_middleware, trace)
 
 from hug import defaults  # isort:skip - must be imported last for defaults to have access to all modules